chore: ProfilePage renamed to Profile
