+++ src/App.js
@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 import { BrowserRouter, Route, Redirect, Switch } from "react-router-dom";
-import ProfilePage from "./ProfilePage";
+import Profile from "./Profile";
 
 export default function App() {
   return (
     <BrowserRouter>
       <div>
         <Redirect from="/" to="/EveryInteract" />
-        <Route exact path="/EveryInteract" component={ProfilePage} />
+        <Route exact path="/EveryInteract" component={Profile} />
       </div>
     </BrowserRouter>
   );
