+++ src/Profile/Content/ProfileInfo/Artefact.js
